+++ src/global.d.ts
@@ -33,10 +33,6 @@ declare module '*.svg' {
 
 declare module '*.vue' {
   import type { DefineComponent } from 'vue';
-  const component: DefineComponent<
-    Record<string, never>,
-    Record<string, never>,
-    any
-  >;
+  const component: DefineComponent<Record<string, never>, Record<string, never>, any>;
   export default component;
 }