+++ src/router/index.ts
@@ -165,7 +165,7 @@ const router = createRouter({
 });
 
 // 路由守卫
-router.beforeEach((to, from) => {
+router.beforeEach((to) => {
   const user = useUserStore();
 
   if (to.name === 'Auth' && user.isLogin) {