+++ src/utils/imageCompression.ts
@@ -16,21 +16,11 @@ const dataURLToImage = (dataURL: string): Promise<HTMLImageElement> => {
   });
 };
 
-const canvastoFile = (
-  canvas: HTMLCanvasElement,
-  type: string,
-  quality: number,
-): Promise<Blob | null> => {
-  return new Promise((resolve) =>
-    canvas.toBlob((blob) => resolve(blob), type, quality),
-  );
+const canvastoFile = (canvas: HTMLCanvasElement, type: string, quality: number): Promise<Blob | null> => {
+  return new Promise((resolve) => canvas.toBlob((blob) => resolve(blob), type, quality));
 };
 
-export const compressionFile = async (
-  file: File,
-  type: string = 'image/jpeg',
-  quality: number = 0.5,
-) => {
+export const compressionFile = async (file: File, type: string = 'image/jpeg', quality: number = 0.5) => {
   const fileName = file.name;
   const canvas = document.createElement('canvas');
   const context = canvas.getContext('2d') as CanvasRenderingContext2D;